[rewrite-app-mongo-logging] refactor logger code
