+++ src/app/api/route.ts
@@ -1,13 +1,11 @@
-import {
-  GetDataCronSingleton,
-  mainDataFetch
-} from '@/server/services/dataManager.service';
+import { mainDataFetch } from '@/server/services/dataManager.service';
+import { logger } from '@/server/utils/logger';
 import { NextResponse } from 'next/server';
 
 export const dynamic = 'force-dynamic';
 
 export async function GET() {
-  GetDataCronSingleton.getInstance();
-  const r = await mainDataFetch();
-  return NextResponse.json({ api: 'online', r });
+  const data = await mainDataFetch();
+  logger.info('api/', { success: data?.success, date: new Date() });
+  return NextResponse.json({ api: 'online', data });
 }
